Upgrade to Typescript
