--- index.js
@@ -1,25 +1,0 @@
-export {
-  Model,
-  EntityBase, 
-  ModelBinder, 
-  modelBootstrap, 
-  ModelBoot, 
-  Entity,
-  ModelSeed, 
-  BinderHelper, 
-  Hook, 
-  EnableDisableRemoteMethods, 
-  GetMethodsFromModel,
-  randomId,
-  RequireObject,
-  isFunction,
-  modelLoader,
-  dataSourceLoader,
-  PropertyListChanged,
-  toSpinalCase,
-  PathJoin,
-  ReadFileSync,
-  ReadGlob,
-  GetModelSchema,
-  GetDSConnector
-} from './src';
