+++ package.json
@@ -2,12 +2,10 @@
   "name": "loopback-model-binder",
   "version": "1.2.2",
   "description": "Loopback Model Binder for ES6 classes",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
-    "prepublish": "npm run build",
-    "copy-index": "gulp copy-index --index '.temp'",
-    "build": "gulp build --build 'dist' && npm run copy-index && gulp compile-index --compile './'",
-    "start": "node .'",
+    "build": "./node_modules/.bin/gulp build",
+    "start": "node dist/index.js'",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -21,6 +19,10 @@
   },
   "homepage": "https://github.com/aelbore/loopback-model-binder#readme",
   "devDependencies": {
+    "@types/glob": "^5.0.30",
+    "@types/gulp": "^4.0.2",
+    "@types/node": "^7.0.8",
+    "@types/rx": "^4.1.1",
     "babel": "^6.5.2",
     "babel-core": "^6.9.0",
     "babel-plugin-transform-async-to-generator": "^6.8.0",
@@ -28,12 +30,11 @@
     "babel-preset-es2015": "^6.9.0",
     "babel-register": "^6.9.0",
     "babel-runtime": "^6.11.6",
-    "del": "^2.2.1",
     "gulp": "^3.9.1",
     "gulp-babel": "^6.1.2",
-    "gulp-param": "^1.0.3",
-    "gulp-sourcemaps": "^1.6.0",
-    "vinyl-paths": "^2.1.0"
+    "gulp-typescript": "^3.1.6",
+    "rimraf": "^2.6.1",
+    "typescript": "^2.2.1"
   },
   "dependencies": {
     "glob": "^7.0.5",
