+++ src/entity-base.ts
@@ -1,6 +1,4 @@
-/// <reference path="../typings/index.d.ts" />
-
-import { Bind, Routes  } from './entity-base-utils';
+import { Bind, Routes } from './entity-base-utils';
 
 let modelName, route;
 
@@ -15,7 +13,7 @@ export default class EntityBase {
    * @param {any} ModelName => name of the Model that you created
    * @param {any} Route => its optional, single route that you created
    */
-  constructor(ModelName, Route){
+  constructor(ModelName, Route) {
     modelName = ModelName;
     route = Route;
   }
@@ -25,19 +23,19 @@ export default class EntityBase {
    * Attach the Entity and Routes/remoteMethod to Model. 
    * @return {void}
    */
-  onInit(){
+  onInit() {
     Routes(route, this.constructor.name)
-      .subscribe((routes) => {
-        for(let i = 0; i < routes.length; i++){
+      .subscribe((routes:any) => {
+        for (let i = 0; i < routes.length; i++) {
           let routeElement = routes[i];
-          if (Array.isArray(routeElement)){
+          if (Array.isArray(routeElement)) {
             routeElement.forEach((element) => {
               Bind(modelName, element, this);
             });
           } else {
             Bind(modelName, routeElement, this);
-          } 
-        } 
+          }
+        }
       });
   }
