+++ src/entity-collections.ts
@@ -1,5 +1,3 @@
-/// <reference path="../typings/index.d.ts" />
-
 let entities = Symbol();
 let singleton = Symbol();
 let _currentModel = Symbol();
@@ -9,19 +7,19 @@ export default class Entity {
     this[entities] = new Array();
   }
 
-  static get collection(){
-    if (!this[singleton]){
+  static get collection() {
+    if (!this[singleton]) {
       this[singleton] = new Entity();
-    } 
-    return this[singleton];    
+    }
+    return this[singleton];
   }
 
-  add(model, entity, routes){
+  add(model, entity, routes) {
     this[_currentModel] = model;
     this[entities].push({ model: model, entity: entity, routes: routes });
   }
 
-  get entities(){
+  get entities() {
     return this[entities];
   }
