+++ src/model-boot-utils.ts
@@ -1,15 +1,13 @@
-/// <reference path="../typings/index.d.ts" />
-
 import { modelLoader } from './model-loader';
 import * as Rx from 'rx';
 
 const IGNORE_FILES = [
-    "./**/*.route.js",
-    "./index.js",
-    "./*-binder.config.js",
-    "./*-boot.js",
-    "./*-datasources.json",
-    "./*-model.json"  
+  "./**/*.route.js",
+  "./index.js",
+  "./*-binder.config.js",
+  "./*-boot.js",
+  "./*-datasources.json",
+  "./*-model.json"
 ];
 
 let modelBootUtils = {
@@ -17,32 +15,32 @@ let modelBootUtils = {
     return Rx.Observable.create((observer) => {
       modelBootUtils
         .load(app, rootDir, dataSources, configs)
-        .subscribe((value) => {
-          modelBootUtils.extend(app, value.dataSources, value.configs);  
+        .subscribe((value: any) => {
+          modelBootUtils.extend(app, value.dataSources, value.configs);
           observer.onNext(value.modelName);
-          observer.onCompleted();   
+          observer.onCompleted();
         });
     });
   },
   load: (app, rootDir, dataSources, configs) => {
     let localConfigs = Object.assign({}, configs);
     localConfigs.dataSource = dataSources[0];
-    localConfigs.rootDir = rootDir; 
-    return modelLoader.load(app, localConfigs) 
+    localConfigs.rootDir = rootDir;
+    return modelLoader.load(app, localConfigs)
       .flatMap((modelName) => {
         return Rx.Observable.create((observer) => {
-          observer.onNext({ 
-            configs: localConfigs,  
+          observer.onNext({
+            configs: localConfigs,
             modelName: modelName,
             dataSources: dataSources
-          }); 
+          });
           observer.onCompleted();
         });
-      });  
+      });
   },
   extend: (app, dataSources, configs) => {
-    if (dataSources.length > 1){
-      for(let i = 1; i < dataSources.length; i++){
+    if (dataSources.length > 1) {
+      for (let i = 1; i < dataSources.length; i++) {
         modelLoader.extends(app, dataSources[i], configs);
       }
     }
