+++ src/model-boot.ts
@@ -1,5 +1,3 @@
-/// <reference path="../typings/index.d.ts" />
-
 import { dataSourceLoader } from './datasource-loader';
 import { modelBootUtils, IGNORE_FILES } from './model-boot-utils';
 import { ReadGlob } from './utils';
@@ -12,13 +10,13 @@ let rootDir = Symbol();
  * let configs = {
  *   files: [
  *     { include: ".//**//*.js", ignore: [] }
- *   ],
- *   seed: {
- *     rootDir: __dirname,
- *     isSeed: true
- *   }
- * }
- */
+*   ],
+*   seed: {
+*     rootDir: __dirname,
+*     isSeed: true
+*   }
+* }
+*/
 
 /**
  * Boots the Model/Datasource that you create
@@ -39,7 +37,7 @@ export default class ModelBoot {
   /**
    * Initialize the datasource and model
    */
-  onInit(){
+  onInit() {
     return dataSourceLoader
       .load(this[app], this[rootDir])
       .flatMap((dataSources) => {
@@ -54,15 +52,15 @@ export default class ModelBoot {
    * by default it will check *-binder.config.js file
    * if not exist will create default config schema
   */
-  get configs(){
+  get configs() {
     let configValues = null;
     let _configs = ReadGlob(`${this[rootDir]}/*-binder.config.js`);
-    if (_configs && _configs.length > 0){
-      if (_configs.length > 1){
-        throw new Error(`Only one (1) model binder.config file, should be in ${this[rootDir]}`);  
-      }    
+    if (_configs && _configs.length > 0) {
+      if (_configs.length > 1) {
+        throw new Error(`Only one (1) model binder.config file, should be in ${this[rootDir]}`);
+      }
       _configs.forEach((configSchema) => {
-        configValues = require(configSchema).configs; 
+        configValues = require(configSchema).configs;
       });
     } else {
       const SEED_ROOT_DIR = `${this[rootDir]}/seeds`;
