+++ src/model-loader-utils.ts
@@ -1,5 +1,3 @@
-/// <reference path="../typings/index.d.ts" />
-
 import { toSpinalCase, ReadGlob, PathJoin, RequireObject, isFunction } from './utils';
 import ModelSeed from './model-seed';
 import * as Rx from 'rx';
@@ -8,24 +6,24 @@ import * as Rx from 'rx';
 ///   It should have Model/Collections list that exempt 
 ///   from seeding data
 let SeedData = (seed, model, collection, dataSource) => {
-  if (seed.isSeed){
+  if (seed.isSeed) {
     let files = ReadGlob(PathJoin(seed.rootDir, `./${toSpinalCase(collection)}.js`));
-    if (files){
+    if (files) {
       let RxNodeCallBack = Rx.Observable.fromNodeCallback(seeFile);
       Rx.Observable
         .for(files, file => RxNodeCallBack(file, model, dataSource))
-        .subscribe(() => {});
+        .subscribe(() => { });
     }
   }
-}, 
-seeFile = (file, model, dataSource) => {
-  let seed = RequireObject(file);
-  if (seed && isFunction(seed)){ 
-    let seedObject = new seed(model, dataSource);
-    if (seedObject instanceof ModelSeed){
-      seedObject.execute(); 
+},
+  seeFile = (file, model, dataSource) => {
+    let seed = RequireObject(file);
+    if (seed && isFunction(seed)) {
+      let seedObject = new seed(model, dataSource);
+      if (seedObject instanceof ModelSeed) {
+        seedObject.execute();
+      }
     }
-  }
-};
+  };
 
 export { SeedData };
