+++ tasks/clean.js
@@ -1,0 +1,6 @@
+const gulp = require('gulp');
+const rimraf = require('rimraf');
+
+gulp.task('clean', (callback) => {
+  return rimraf('dist', callback);
+});
