--- typings.json
@@ -1,15 +1,0 @@
-{
-  "name": "loopback-model-binder",
-  "dependencies": {},
-  "globalDependencies": {
-    "async": "registry:dt/async#2.0.1+20160804113311",
-    "es6-shim": "registry:dt/es6-shim#0.31.2+20160602141504",
-    "glob": "registry:dt/glob#5.0.10+20160317120654",
-    "gulp": "registry:dt/gulp#3.8.0+20160316155526",
-    "gulp-babel": "registry:dt/gulp-babel#6.1.0+20160316171810",
-    "gulp-sourcemaps": "registry:dt/gulp-sourcemaps#0.0.0+20160316155526",
-    "node": "registry:dt/node#6.0.0+20160709063457",
-    "rx": "registry:dt/rx#0.0.0+20160317120654",
-    "rx-node": "registry:dt/rx-node#0.0.0+20160513125029"
-  }
-}
